requirements.txt
Flask
openai
newspaper3k
requests
python-dotenv
flask-cors
